requirements.txt
asgiref==3.6.0
certifi==2023.5.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.16
django-cleanup==7.0.0
django-debug-toolbar==4.0.0
django-filter==23.1
django-phonenumber-field==7.1.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
install==1.3.5
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
openapi-codec==1.3.2
phonenumbers==8.13.11
phonenumberslite==8.13.11
Pillow==9.5.0
PyJWT==2.7.0
pytz==2022.7.1
requests==2.31.0
rest-framework-simplejwt==0.0.2
routers==0.10.1
simplejson==3.19.1
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==2.0.3
